+++ Changelog.md
@@ -4,7 +4,7 @@
 - Fix action "Add node pool"
 - Change default value for "Inherit DSS host settings" when creating a cluster
 - Change default value for "Service account type" in the nodes pool creation form
-- Handle autopilot clusters
+- Handle regional clusters
 
 ## Version 1.1.4 - Feature release
 - Add ability to use named secondary ranges