+++ Changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next version
+## Version 1.2.0
 - Fix the cluster configuration's merging mechanism for string parameters
 - Remove support for legacy authorization (ABAC)
 - Authentication for `kubectl` command now relies on `gke-gcloud-auth-plugin` as this is the only supported mode starting with Kubernetes 1.26 (see [documentation](https://cloud.google.com/blog/products/containers-kubernetes/kubectl-auth-changes-in-gke))